Midway through making proxy settings available to httpeasy and adding tests
