+++ cubano-config/build.gradle
@@ -1,0 +1,6 @@
+
+dependencies {
+    
+    // compile 'org.powermock:powermock:1.6.6'
+    testCompile 'org.powermock:powermock-core:1.7.3'
+}
