+++ cubano-config/src/test/java/org/concordion/cubano/utils/ConfigMock.java
@@ -1,0 +1,15 @@
+package org.concordion.cubano.utils;
+
+import java.util.Properties;
+
+public class ConfigMock extends Config {
+
+	public ConfigMock(Properties properties) {
+		super(properties);
+	}
+	
+	@Override
+	protected void loadProperties() {
+		
+	}
+}
