+++ cubano-httpeasy/build.gradle
@@ -1,6 +1,9 @@
 dependencies {
     compile project(':cubano-http')
-
+    
+	// TODO Ideally httpeasy wouldn't dependencies on other components to make it easier to use outside of cubano
+	compile project(':cubano-config')
+	
     compile group: 'com.google.guava', name: 'guava', version: '22.0'
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.1'
 }
